+++ frontend/src/App.tsx
@@ -30,9 +30,8 @@ import { USER_CLOSE_WEB_SOCKET_CODE } from "./constants";
 import CodeTab from "./components/CodeTab";
 import OutputSettingsSection from "./components/OutputSettingsSection";
 import { History } from "./history_types";
-import HistoryDisplay, {
-  extractHistoryTree,
-} from "./components/HistoryDisplay";
+import HistoryDisplay from "./components/history/HistoryDisplay";
+import { extractHistoryTree } from "./components/history/utils";
 import toast from "react-hot-toast";
 
 const IS_OPENAI_DOWN = false;
